// src/app/models/tournament.ts
export interface Tournament {
    uid: string;
    userID: string;
    tournamentName?: string | null;
    startDate?: string | null;
    endDate?: string | null;
    startTime?: string | null;
    clubName?: string | null;
    city?: string | null;
    state?: string | null;
    address?: string | null;
    addressNumber?: string | null;
    addressComp?: string | null;
    zipCode?: string | null;
    photoURL?: string | null;
  }
  
  
  